# docs/api/quant-radar-v1.openapi.yaml
openapi: 3.1.0
info:
  title: Surge Screener API
  version: 1.23.0-draft
  summary: Loopback-only public reads plus protected private operator state.
  description: |
    Contract for the current Phase 1 fixed-artifact reads plus bounded Phase 2
    reads for IV history, the institutional fund catalog, the manually maintained
    AI updates feed, the manually maintained schedule registry, a strict
    Options Flow and candidate feed projections, the strict Crypto Universe,
    Social Intelligence, Theme Flow, Money Flow, Sector Rotation, Daily Summary,
    Playbook Validation, Continuation Validation, COT report catalog/detail, and
    bounded Market Thesis validation/regime-history, Oversold Reversal
    validation, Reversal Radar validation, Knowledge Graph, Watchlist theme
    taxonomy, sector-to-theme drill, and Influencer roster public projections.
    Expected
    artifact failures are represented as HTTP 200 unavailable envelopes so a
    missing or half-written scheduled output cannot crash the frontend. Invalid
    request parameters use sanitized HTTP 422 problem details. Unexpected server
    defects remain HTTP 500.

    Public artifact routes intentionally have no authentication scheme because
    the server is restricted to 127.0.0.1 and exposes bounded non-sensitive
    projections. The private Industry Roles review-board routes additionally
    require a high-entropy Streamlit-to-FastAPI bearer credential. Mutations
...
servers:
  - url: http://127.0.0.1:8000
...
security: []
...
paths:
...
  /api/v1/market-context/money-flow/latest:
    get:
      tags: [Market Context]
      summary: Read the latest strict Money Flow snapshot
      description: |
        Returns the bounded allowlisted Eastmoney presentation projection.
        Provider identifiers, raw rows, and unused flow breakdowns remain
        server-side.
      operationId: getLatestMoneyFlow
      responses:
        "200":
          $ref: "#/components/responses/MoneyFlowResponse"
        "500":
          $ref: "#/components/responses/InternalProblem"
...
components:
  securitySchemes:
    InternalServiceBearer:
      type: http
      scheme: bearer
      description: |
        High-entropy Streamlit-to-FastAPI service credential for protected
        loopback routes; not a human login or provider token.
  headers:
    NoStore:
      description: Prevent browsers and intermediaries from retaining artifact state.
      schema:
        type: string
        const: no-store
...
  responses:
...
    MoneyFlowResponse:
      description: Strict persisted Money Flow presentation state.
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MoneyFlowEnvelope"
...
    InternalProblem:
      description: Unexpected programming or server failure.
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: about:blank
            title: Internal Server Error
            status: 500
  schemas:
...
    AvailableArtifactBase:
      type: object
      required: [available, reason, meta]
      properties:
        available:
          type: boolean
          const: true
        reason:
          type: string
          const: ok
        meta:
          $ref: "#/components/schemas/ArtifactMeta"
...
    AvailableMoneyFlow:
      unevaluatedProperties: false
      allOf:
        - $ref: "#/components/schemas/AvailableArtifactBase"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/MoneyFlowData"
...
    UnavailableArtifact:
      type: object
      additionalProperties: false
      required: [available, reason, data, meta]
      properties:
        available:
          type: boolean
          const: false
        reason:
          $ref: "#/components/schemas/UnavailableReason"
        data:
          type: "null"
        meta:
          $ref: "#/components/schemas/ArtifactMeta"
    UnavailableReason:
      type: string
      enum:
        - missing
        - invalid_json
        - invalid_shape
        - unreadable
...
    MoneyFlowEnvelope:
      oneOf:
        - $ref: "#/components/schemas/AvailableMoneyFlow"
        - $ref: "#/components/schemas/UnavailableArtifact"
...
    MoneyFlowData:
      type: object
      additionalProperties: false
      description: Strict Money Flow presentation projection.
      required:
        - as_of_date
        - generated_at
        - source
        - publishable
        - coverage
        - rows
      properties:
        as_of_date:
          type: string
          format: date
          pattern: '^\d{4}-\d{2}-\d{2}(?![\s\S])'
        generated_at:
          type: string
          format: date-time
          minLength: 20
          maxLength: 32
          pattern: '^\d{4}-\d{2}-\d{2}[Tt](?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d{1,6})?(?:[Zz]|[+-](?:[01]\d|2[0-3]):[0-5]\d)(?![\s\S])'
        source:
          type: string
          const: eastmoney_push2his
        publishable:
          type: boolean
        coverage:
          $ref: "#/components/schemas/MoneyFlowCoverage"
        rows:
          type: array
          maxItems: 50000
          items:
            $ref: "#/components/schemas/MoneyFlowRow"
    MoneyFlowCoverage:
      type: object
      additionalProperties: false
      description: Closed producer coverage accounting for one Money Flow snapshot.
      required: [requested, resolved, unavailable, coverage_ratio, min_coverage]
      properties:
        requested:
          type: integer
          minimum: 0
          maximum: 100000
        resolved:
          type: integer
          minimum: 0
          maximum: 100000
        unavailable:
          type: integer
          minimum: 0
          maximum: 100000
        coverage_ratio:
          type: number
          minimum: 0
          maximum: 1
        min_coverage:
          type: number
          minimum: 0
          maximum: 1
    MoneyFlowRow:
      type: object
      additionalProperties: false
      description: Allowlisted daily flow fields used by approved presentation consumers.
      required: [ticker, date, main_net, main_pct, small_net, source]
      properties:
        ticker:
          type: string
          minLength: 1
          maxLength: 15
          pattern: '^[A-Z0-9]+(?:[.-][A-Z0-9]+)*(?![\s\S])'
        date:
          type: string
          format: date
          pattern: '^\d{4}-\d{2}-\d{2}(?![\s\S])'
        main_net:
          anyOf:
            - type: number
              minimum: -1000000000000000
              maximum: 1000000000000000
            - type: "null"
        main_pct:
          anyOf:
            - type: number
              minimum: -1000000
              maximum: 1000000
            - type: "null"
        small_net:
          anyOf:
            - type: number
              minimum: -1000000000000000
              maximum: 1000000000000000
            - type: "null"
        source:
          type: string
          const: eastmoney_push2his
...
    ArtifactMeta:
      type: object
      additionalProperties: false
      required: [sourceId, asOf, generatedAt]
      properties:
        sourceId:
          type: string
          description: Stable registry identifier; never a filesystem path.
          pattern: "^[a-z0-9]+(?:[.-][a-z0-9]+)*$"
        asOf:
          type: [string, "null"]
          format: date
        generatedAt:
          type: [string, "null"]
          format: date-time
    Problem:
      type: object
      description: RFC 9457 problem details for unexpected server failures.
      additionalProperties: true
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference